requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
idna==3.4
joblib==1.2.0
keyboard==0.13.5
loguru==0.6.0
MouseInfo==0.1.3
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
PyAutoGUI==0.9.44
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.10
python3-xlib==0.15
pytweening==1.0.4
regex==2023.3.23
requests==2.28.2
sydney-py==0.9.0
tqdm==4.65.0
urllib3==1.26.15
websockets==10.4
yarl==1.8.2
